Remove embedded libmbedtls on Debian
https://github.com/bibledit/cloud/issues/168

+++ client-debian/tarball.sh
@@ -23,6 +23,11 @@ DEBIANSOURCE=`pwd`
 echo Using Debian packaging source at $DEBIANSOURCE.
 
 
+echo Create a tarball for the Linux Client
+../../linux/tarball.sh
+if [ $? -ne 0 ]; then exit; fi
+
+
 TMPDEBIAN=/tmp/bibledit-debian
 echo Working folder $TMPDEBIAN
 rm -rf $TMPDEBIAN
@@ -31,7 +36,7 @@ cd $TMPDEBIAN
 if [ $? -ne 0 ]; then exit; fi
 
 
-# The script unpacks the Bibledit Linux tarball,
+# The script unpacks the Bibledit Linux tarball created above,
 # modifies it, and repacks it into a Debian tarball.
 # The reason for doing so is that the Debian builder would otherwise notice
 # differences between the supplied tarball and the modified source.
@@ -39,7 +44,7 @@ if [ $? -ne 0 ]; then exit; fi
 # Another reason is that in this way it does not need to generate patches in the 'debian' folder.
 
 
-echo Unpack the tarball assumed to be created for the Bibledit Linux client.
+echo Unpack the Linux client tarball created just now
 tar xf ~/Desktop/bibledit*gz
 if [ $? -ne 0 ]; then exit; fi
 cd bibledit*
@@ -50,15 +55,22 @@ echo Copy the Debian packaging source to $TMPDEBIAN
 cp -r $DEBIANSOURCE/debian .
 
 
-# echo Link with the system-provided mbed TLS library.
+echo Link with the system-provided mbed TLS library.
+# It is important to use the system-provided mbedtls library because it is a security library.
+# This way, Debian updates to libmbedtls become available to Bibledit too.
+# With the embedded library, this is not the case.
 # Fix for lintian error "embedded-library usr/bin/bibledit: mbedtls":
 # * Remove mbedtls from the list of sources to compile.
 # * Add -lmbedtls and friends to the linker flags.
-# sed -i.bak '/mbedtls\//d' Makefile.am
-# if [ $? -ne 0 ]; then exit; fi
-# sed -i.bak 's/# debian//g' Makefile.am
-# if [ $? -ne 0 ]; then exit; fi
-# rm *.bak
+sed -i.bak '/mbedtls\//d' Makefile.am
+if [ $? -ne 0 ]; then exit; fi
+sed -i.bak 's/# debian//g' Makefile.am
+if [ $? -ne 0 ]; then exit; fi
+rm *.bak
+# Also remove the embedded *.h files to be sure building does not reference them.
+# There had been a case that building used the embedded *.h files, leading to segmentation faults.
+# For cleanness, remove the whole mbedtls directory, so all traces of it are gone completely.
+rm -rf mbedtls
 
 
 # If the debian/README* or README.Debian files contain no useful content,
@@ -102,7 +114,7 @@ find . -name .DS_Store -delete
 echo Remove macOS extended attributes.
 # The attributes would make their way into the tarball,
 # get unpacked within Debian,
-# and would cause build errors.
+# and would cause build errors there.
 xattr -r -c *
